requirements.txt
#Core
python-dotenv

#Scrapping
praw

#Screenshots
playwright
validators

#Voice 
gTTS
requests
playsound==1.2.2
gradio_client

#Video Editing
moviepy

#Captions
assemblyai
pysrt